move delta back
